metric bug fix

+++ ryan/training.py
@@ -65,7 +65,7 @@ def train(model_name, input_path, output_file, epochs, batch_size, val_split, ex
         return
 
     def exact_pred(y_true, y_pred):
-        return K.min(K.equal(y_true, K.round(y_pred)), axis=-1)
+        return K.min(K.cast(K.equal(y_true, K.round(y_pred)), dtype='float16'), axis=-1)
 
     # Compile model
     model.compile(loss='binary_crossentropy',
